+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change Log
 
+## [0.0.15] - 2025-04-02
+- Chnaged hotkeys From CMD+1/2/3 to 1/2/3 to activate Pen/Shapes/Laser
+- Added hotkeys 4/5 to switch Color and Thickness
+
 ## [0.0.14] - 2025-03-16
 - Added hotkeys CMD+1/2/3 to activate Pen/Shapes/Laser